requirements.txt
django==1.8.3
django-crispy-forms
django-allauth
django-dirtyfields
stripe
docutils
#django-extensions
#django-uni-form
#requests
